refactor(install): split pipeline modules
